requirements.txt
pandas
numpy
scikit-learn
pykan
matplotlib
torch
tqdm
tabulate
pyyaml
loguru
tensorboard